refactor: delete services/ package entirely + fix remaining stale imports
services/ is now empty after all modules migrated to system_services/
or bricks/. Remove the last __init__.py and fix remaining stale refs:

- search_service.py: services.gateway → system_services.gateway
- test_overlay_nexusfs.py: services.workspace → bricks.workspace
- test_async_wrappers_e2e.py: services.routing → core.async_router
- pyproject.toml: update 3 import-linter ratchet entries
